requirements.txt
langchain
langchain-community
langchain-groq
fastapi
httpx
PyMuPDF
streamlit
python-dotenv